+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Fix `useCache` hook when it has to encode UTF-16 characters to base 64.
 - Fix an issue related to css selector priority from r-scheme-colors mixins
 - Fix css classes related with course runs on the course detail page.
+- Fix error on course page when course run hasn't a title and the setting
+WEB_ANALYTICS_ID is enabled
 
 ## Changed